refactor: couchbase external system api
