+++ core/target-systems/couchbase-external-system-api/build.gradle.kts
@@ -1,11 +1,11 @@
 dependencies {
     implementation(project(":core:flamingock-core-api"))
 
-    // MongoDB driver for storage implementations
-    compileOnly("org.mongodb:mongodb-driver-sync:4.0.0")
+    //General
+    compileOnly("com.couchbase.client:java-client:3.6.0")
 }
 
-description = "MongoDB external system api"
+description = "Couchbase external system api"
 
 java {
     toolchain {
