+++ core/target-systems/couchbase-external-system-api/src/main/java/io/flamingock/externalsystem/couchbase/api/CouchbaseExternalSystem.java
@@ -1,4 +1,4 @@
-package io.flamingock.targetsystem.couchbase;
+package io.flamingock.externalsystem.couchbase.api;
 
 import com.couchbase.client.java.Bucket;
 import com.couchbase.client.java.Cluster;
