+++ core/target-systems/mongodb-external-system-api/src/main/java/io/flamingock/externalsystem/mongodb/api/MongoDBExternalSystem.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.flamingock.targetsystem.mongodb.api;
+package io.flamingock.externalsystem.mongodb.api;
 
 import com.mongodb.client.MongoDatabase;
 import io.flamingock.api.external.ExternalSystem;
